ml_models/eeg_classifier/requirements.txt
# EEG Classifier Dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pillow>=9.0.0
pydantic>=2.0.0

# For EEG signal processing
# scipy>=1.9.0
# mne>=1.0.0
